Rename epic files to satisfy SonarCloud
